subservers: move permission logic
