+++ perms/manager_test.go
